Extract EvaluatorScoreCell + readEvaluatorCell into shared module
STTResultsTable and TTSResultsTable each carried byte-identical copies
of the per-row evaluator cell renderer (~90 lines, including the
error / Pass-Fail / score-or-score-max branches) and the helper that
prefers the canonical evaluator_outputs[uuid] shape over the legacy
flat keys. Same logic, modulo the column-type Pick.

Move both to src/components/eval-details/EvaluatorScoreCell.tsx
exporting EvaluatorScoreCell, readEvaluatorCell, and a structural
EvaluatorColumnLike type that both column types satisfy. Both tables
now import from there; LLMJudgeBadge (the legacy single-judge column
renderer) stays per-table since it has different signatures.
